Update README.md
In this project I used the following AWS services, S3, Cloud 9 IDE using an EC2 instance and Rekognition. Here is an example of the architecture. Amazon Rekognition falls under the AI services of AWS, simply put enter an image and Rekognition will give you a prediction of what it is